+++ sources/src/main/java/com/google/solutions/jitaccess/core/RoleBinding.java
@@ -19,10 +19,9 @@
 // under the License.
 //
 
-package com.google.solutions.jitaccess.core.catalog;
+package com.google.solutions.jitaccess.core;
 
 import com.google.common.base.Preconditions;
-import com.google.solutions.jitaccess.core.ProjectId;
 
 import java.util.Comparator;
