+++ sources/src/main/java/com/google/solutions/jitaccess/core/catalog/project/ProjectRoleId.java
@@ -24,7 +24,7 @@ package com.google.solutions.jitaccess.core.catalog.project;
 import com.google.common.base.Preconditions;
 import com.google.solutions.jitaccess.core.ProjectId;
 import com.google.solutions.jitaccess.core.catalog.EntitlementId;
-import com.google.solutions.jitaccess.core.catalog.RoleBinding;
+import com.google.solutions.jitaccess.core.RoleBinding;
 
 public class ProjectRoleId extends EntitlementId {
   static final String CATALOG = "iam";
