+++ sources/src/test/java/com/google/solutions/jitaccess/core/TestRoleBinding.java
@@ -19,8 +19,9 @@
 // under the License.
 //
 
-package com.google.solutions.jitaccess.core.catalog;
+package com.google.solutions.jitaccess.core;
 
+import com.google.solutions.jitaccess.core.RoleBinding;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
