+++ sources/src/test/java/com/google/solutions/jitaccess/core/catalog/project/TestIamPolicyCatalog.java
@@ -21,10 +21,7 @@
 
 package com.google.solutions.jitaccess.core.catalog.project;
 
-import com.google.solutions.jitaccess.core.AccessDeniedException;
-import com.google.solutions.jitaccess.core.Annotated;
-import com.google.solutions.jitaccess.core.ProjectId;
-import com.google.solutions.jitaccess.core.UserId;
+import com.google.solutions.jitaccess.core.*;
 import com.google.solutions.jitaccess.core.catalog.*;
 import com.google.solutions.jitaccess.core.clients.ResourceManagerClient;
 import org.junit.jupiter.api.Test;
