+++ sources/src/test/java/com/google/solutions/jitaccess/web/rest/TestApiResource.java
@@ -22,10 +22,7 @@
 package com.google.solutions.jitaccess.web.rest;
 
 import com.google.auth.oauth2.TokenVerifier;
-import com.google.solutions.jitaccess.core.AccessDeniedException;
-import com.google.solutions.jitaccess.core.Annotated;
-import com.google.solutions.jitaccess.core.ProjectId;
-import com.google.solutions.jitaccess.core.UserId;
+import com.google.solutions.jitaccess.core.*;
 import com.google.solutions.jitaccess.core.catalog.*;
 import com.google.solutions.jitaccess.core.catalog.project.IamPolicyCatalog;
 import com.google.solutions.jitaccess.core.catalog.project.ProjectRoleActivator;
